requirements.txt
alembic==1.12.0
asyncpg==0.28.0
fastapi==0.103.2
greenlet==3.0.0
httpx==0.25.0
pydantic==2.4.2
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlalchemy==2.0.21
uvicorn==0.23.2